+++ age-core/CHANGELOG.md
@@ -8,9 +8,14 @@ to 1.0.0 are beta releases.
 
 ## [Unreleased]
 ### Added
-- `age_core::format::is_arbitrary_string`
+- `age_core::format`:
+  - `FileKey::new`
+  - `FileKey::init_with_mut`
+  - `FileKey::try_init_with_mut`
+  - `is_arbitrary_string`
 
 ### Changed
+- Migrated to `secrecy 0.10`.
 - `age::plugin::Connection::unidir_receive` now takes an additional argument to
   enable handling an optional fourth command.